requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==9.4.0
joblib==1.2.0
timm==0.9.8
datasets==2.19.1
numpy==1.24.4
scipy==1.10.1
scikit-learn==1.3.2
secml-torch
foolbox
tensorboard
fairscale
matplotlib
transformers==4.17.0
git+https://github.com/openai/CLIP.git
albumentations==2.0.5
tensorboardX
faiss-gpu==1.7.2
scikit-image
diffusers